[refactor] Modulization (HOME)
HomeBanner와 Slider로 컴포넌트 분리
