+++ vite-project/src/index.css
@@ -15,6 +15,12 @@
   -moz-osx-font-smoothing: grayscale;
 }
 
+* {
+  box-sizing: border-box;
+  margin: 0;
+  padding: 0;
+}
+
 html, body, #root {
   margin: 0;
   padding: 0;
@@ -34,6 +40,8 @@ body {
   place-items: center;
   min-width: 320px;
   min-height: 100vh;
+  font-family: sans-serif;
+  color: black;
 }
 
 h1 {
